docker: enable more caching

+++ docker/util.sh
@@ -17,7 +17,7 @@ fi
 
 usage()
 {
-    echo "Usage: build.sh [--push] [--docker-repository repo] [--tag name] [--gdal tag|sha1|master] [--gdal-repository repo] [--proj tag|sha1|master] [--release]"
+    echo "Usage: build.sh [--push] [--docker-repository repo] [--tag name] [--gdal tag|sha1|master] [--gdal-repository repo] [--proj tag|sha1|master] [--release] [--docker-cache|--no-docker-cache] [--no-rsync-daemon]"
     # Non-documented: --test-python
     echo ""
     echo "--push: push image to docker repository (defaults to ghcr.io)"
@@ -25,7 +25,9 @@ usage()
     echo "--gdal tag|sha1|master: GDAL version to use. Defaults to master"
     echo "--proj tag|sha1|master: PROJ version to use. Defaults to master"
     echo "--gdal-repository repo: github repository. Defaults to OSGeo/gdal"
-    echo "--release. Whether this is a release build. In which case --gdal tag must be used."
+    echo "--release: Whether this is a release build. In which case --gdal tag must be used."
+    echo "--docker-cache/--no-docker-cache: instruct Docker to build with/without using its cache. Defaults to no cache for release builds."
+    echo "--no-rsync-daemon: do not use the rsync daemon to save build cache in home directory."
     echo "--with-debug-symbols/--without-debug-symbols. Whether to include debug symbols. Only applies to ubuntu-full, default is to include for non-release builds."
     exit 1
 }
@@ -100,6 +102,20 @@ do
             shift
         ;;
 
+        --docker-cache)
+            DOCKER_CACHE_PARAM=""
+            shift
+        ;;
+        --no-docker-cache)
+            DOCKER_CACHE_PARAM="--no-cache"
+            shift
+        ;;
+
+        --no-rsync-daemon)
+            NO_RSYNC_DAEMON=1
+            shift
+        ;;
+
         --without-debug-symbols)
             WITH_DEBUG_SYMBOLS=no
             shift
@@ -145,6 +161,7 @@ if test "${RELEASE}" = "yes"; then
     if test "${WITH_DEBUG_SYMBOLS}" = ""; then
         WITH_DEBUG_SYMBOLS=no
     fi
+    [ -v DOCKER_CACHE_PARAM ] || DOCKER_CACHE_PARAM="--no-cache"
 else
     if test "${TAG_NAME}" = ""; then
         TAG_NAME=latest
@@ -152,6 +169,7 @@ else
     if test "${WITH_DEBUG_SYMBOLS}" = ""; then
         WITH_DEBUG_SYMBOLS=yes
     fi
+    [ -v DOCKER_CACHE_PARAM ] || DOCKER_CACHE_PARAM=""
 fi
 
 check_image()
@@ -194,15 +212,17 @@ echo "Using GDAL_REPOSITORY=${GDAL_REPOSITORY}"
 IMAGE_NAME="${TARGET_IMAGE}-${TAG_NAME}"
 REPO_IMAGE_NAME="${DOCKER_REPO}/${IMAGE_NAME}"
 
+BUILD_ARGS=(
+    "--build-arg" "PROJ_DATUMGRID_LATEST_LAST_MODIFIED=${PROJ_DATUMGRID_LATEST_LAST_MODIFIED}" \
+    "--build-arg" "PROJ_VERSION=${PROJ_VERSION}" \
+    "--build-arg" "GDAL_VERSION=${GDAL_VERSION}" \
+    "--build-arg" "GDAL_REPOSITORY=${GDAL_REPOSITORY}" \
+    "--build-arg" "WITH_DEBUG_SYMBOLS=${WITH_DEBUG_SYMBOLS}" \
+)
+[ -z "${DOCKER_CACHE_PARAM}" ] || BUILD_ARGS+=("${DOCKER_CACHE_PARAM}")
+
 if test "${RELEASE}" = "yes"; then
-    BUILD_ARGS=(
-        "--build-arg" "PROJ_DATUMGRID_LATEST_LAST_MODIFIED=${PROJ_DATUMGRID_LATEST_LAST_MODIFIED}" \
-        "--build-arg" "PROJ_VERSION=${PROJ_VERSION}" \
-        "--build-arg" "GDAL_VERSION=${GDAL_VERSION}" \
-        "--build-arg" "GDAL_REPOSITORY=${GDAL_REPOSITORY}" \
-        "--build-arg" "GDAL_BUILD_IS_RELEASE=YES" \
-        "--build-arg" "WITH_DEBUG_SYMBOLS=${WITH_DEBUG_SYMBOLS}" \
-    )
+    BUILD_ARGS+=("--build-arg" "GDAL_BUILD_IS_RELEASE=YES")
 
     if test "${BASE_IMAGE}" != ""; then
         BUILD_ARGS+=("--build-arg" "BASE_IMAGE=${BASE_IMAGE}")
@@ -236,6 +256,8 @@ if test "${RELEASE}" = "yes"; then
     fi
 
 else
+    BUILD_ARGS+=("--build-arg" "BUILDKIT_INLINE_CACHE=1")
+    BUILD_ARGS+=("--build-arg" "WITH_CCACHE=1")
 
     IMAGE_NAME_WITH_ARCH="${REPO_IMAGE_NAME}"
     if test "${IMAGE_NAME}" = "osgeo/gdal:ubuntu-full-latest" \
@@ -254,21 +276,23 @@ else
         GDAL_RELEASE_DATE=$(date "+%Y%m%d")
     fi
     echo "Using GDAL_RELEASE_DATE=${GDAL_RELEASE_DATE}"
+    BUILD_ARGS+=("--build-arg" "GDAL_RELEASE_DATE=${GDAL_RELEASE_DATE}")
 
-    RSYNC_DAEMON_CONTAINER=gdal_rsync_daemon
-    HOST_CACHE_DIR="$HOME/gdal-docker-cache"
+    if [ ! -v NO_RSYNC_DAEMON ]; then
+        RSYNC_DAEMON_CONTAINER=gdal_rsync_daemon
+        HOST_CACHE_DIR="$HOME/gdal-docker-cache"
 
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/proj/x86_64"
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/proj/aarch64"
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/gdal/x86_64"
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/gdal/aarch64"
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/spatialite"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/proj/x86_64"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/proj/aarch64"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/gdal/x86_64"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/gdal/aarch64"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/spatialite"
 
-    # Start a Docker container that has a rsync daemon, mounting HOST_CACHE_DIR
-    if ! docker ps | grep "${RSYNC_DAEMON_CONTAINER}"; then
-        RSYNC_DAEMON_IMAGE=osgeo/gdal:gdal_rsync_daemon
-        docker rmi "${RSYNC_DAEMON_IMAGE}" 2>/dev/null || true
-        docker $(build_cmd) -t "${RSYNC_DAEMON_IMAGE}" - <<EOF
+        # Start a Docker container that has a rsync daemon, mounting HOST_CACHE_DIR
+        if ! docker ps | grep "${RSYNC_DAEMON_CONTAINER}"; then
+            RSYNC_DAEMON_IMAGE=osgeo/gdal:gdal_rsync_daemon
+            docker rmi "${RSYNC_DAEMON_IMAGE}" 2>/dev/null || true
+            docker $(build_cmd) -t "${RSYNC_DAEMON_IMAGE}" - <<EOF
 FROM alpine
 
 VOLUME /opt/gdal-docker-cache
@@ -286,29 +310,21 @@ CMD rsync --daemon --port 23985 && while sleep 1; do true; done
 
 EOF
 
-        THE_UID=$(id -u "${USER}")
-        THE_GID=$(id -g "${USER}")
+            THE_UID=$(id -u "${USER}")
+            THE_GID=$(id -g "${USER}")
 
-        docker run -d -u "${THE_UID}:${THE_GID}" --rm \
-            -v "${HOST_CACHE_DIR}":/opt/gdal-docker-cache \
-            --name "${RSYNC_DAEMON_CONTAINER}" \
-            --network host \
-            "${RSYNC_DAEMON_IMAGE}"
-
+            docker run -d -u "${THE_UID}:${THE_GID}" --rm \
+                   -v "${HOST_CACHE_DIR}":/opt/gdal-docker-cache \
+                   --name "${RSYNC_DAEMON_CONTAINER}" \
+                   --network host \
+                   "${RSYNC_DAEMON_IMAGE}"
+        fi
+        RSYNC_REMOTE="rsync://127.0.0.1:23985/gdal-docker-cache/${TARGET_IMAGE}"
+        BUILD_ARGS+=(
+            "--build-arg" "RSYNC_REMOTE=${RSYNC_REMOTE}" \
+            "--network" "host" \
+        )
     fi
-
-    RSYNC_REMOTE="rsync://127.0.0.1:23985/gdal-docker-cache/${TARGET_IMAGE}"
-
-    BUILD_ARGS=(
-        "--build-arg" "PROJ_DATUMGRID_LATEST_LAST_MODIFIED=${PROJ_DATUMGRID_LATEST_LAST_MODIFIED}" \
-        "--build-arg" "PROJ_VERSION=${PROJ_VERSION}" \
-        "--build-arg" "GDAL_VERSION=${GDAL_VERSION}" \
-        "--build-arg" "GDAL_REPOSITORY=${GDAL_REPOSITORY}" \
-        "--build-arg" "GDAL_RELEASE_DATE=${GDAL_RELEASE_DATE}" \
-        "--build-arg" "RSYNC_REMOTE=${RSYNC_REMOTE}" \
-        "--build-arg" "WITH_DEBUG_SYMBOLS=${WITH_DEBUG_SYMBOLS}" \
-        "--build-arg" "BUILDKIT_INLINE_CACHE=1" \
-    )
 
     if test "${BASE_IMAGE}" != ""; then
         BUILD_ARGS+=("--build-arg" "BASE_IMAGE=${BASE_IMAGE}")
@@ -349,7 +365,7 @@ EOF
       fi
     fi
 
-    docker $(build_cmd) --network=host "${BUILD_ARGS[@]}" -t "${IMAGE_NAME_WITH_ARCH}" "${SCRIPT_DIR}"
+    docker $(build_cmd) "${BUILD_ARGS[@]}" -t "${IMAGE_NAME_WITH_ARCH}" "${SCRIPT_DIR}"
 
     if test "${DOCKER_BUILDX}" != "buildx"; then
         check_image "${IMAGE_NAME_WITH_ARCH}"
